+++ packages/server-ai/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xpert-ai/server-ai
 
+## 3.9.32
+
+### Patch Changes
+
+- Updated dependencies [a86016e]
+    - @xpert-ai/plugin-sdk@3.18.1
+    - @xpert-ai/server-core@3.9.47
+
 ## 3.9.31
 
 ### Patch Changes